well, but packets input without queue

+++ capnp/types.capnp
@@ -3,7 +3,7 @@
 struct Vector
 {
 	# this is float16
-	x @0: UInt16;
-	y @1: UInt16;
-	z @2: UInt16;
+	x @0: Float32;
+	y @1: Float32;
+	z @2: Float32;
 }
